Generalize java-smell-detector.js and some clean up
