+++ lib/semantic-sniffer.js
@@ -11,7 +11,7 @@ import SmellDetector from './detector/smell-detector'
 const Parser = require('tree-sitter');
 const Java = require('tree-sitter-java');
 
-const {Query, QueryCursor} = Parser
+const {Query} = Parser
 const parser = new Parser();
 /*
 https://flight-manual.atom.io/api/v1.53.0/DisplayMarkerLayer/
@@ -33,15 +33,6 @@ export default {
         '.source.java': Java,
     },
     activate(state) {
-        // TODO: add configuration that allows you to disable rescanning on change
-
-        // this.semanticSnifferView = new SemanticSnifferView(
-        //   state.semanticSnifferViewState
-        // );
-        // this.modalPanel = atom.workspace.addModalPanel({
-        //   item : this.semanticSnifferView.getElement(),
-        //   visible : false
-        // });
 
         // Events subscribed to in atom's system can be easily cleaned up with a CompositeDisposable
         this.subscriptions = new CompositeDisposable();
@@ -56,13 +47,18 @@ export default {
         // https://flight-manual.atom.io/hacking-atom/sections/package-word-count/
         atom.workspace.observeTextEditors(editor => {
             var sniffviewlayer = editor.addMarkerLayer();
+            let lang = snifferLangs[editor.getRootScopeDescriptor()];
+            if(lang != undefined) {
 
-            this.editorMarkerLayers[editor.id] = sniffviewlayer;
-            this.workingTree[editor.id] = this.sniff(editor);
-            // TODO: subscribe to changes to text
-            this.subscriptions.add(editor.onDidStopChanging(event => {
-                this.sniff(editor);
-            }));
+              this.editorMarkerLayers[editor.id] = sniffviewlayer;
+              // this.workingTree[editor.id] = this.sniff(editor);
+              this.sniffers[editor.id] = new SmellDetector(
+                  snifferLangs[lang],
+              );
+              this.subscriptions.add(editor.onDidStopChanging(event => {
+                  this.sniff(editor);
+              }));
+          }
         });
     },
 
@@ -93,15 +89,9 @@ export default {
 
         console.log("Observing editor: " + editor.getLongTitle());
         console.log("Language: " + lang);
-        // bruh bruh
-        var regex_if = new RegExp('(if.*).\\s*(?!==)\\b(?=\\s*=)\\s*=\\s*.*$', 'gm');
-        // var regex_if = new RegExp('bro', 'gm');
-        // for now, if we encounter an if statement, print something to the console.
-        // called for each scan result
 
         let buffer = editor.getBuffer().buffer;
         // can we assume the above buffer updates with editor?
-        const tree = parser.parseTextBufferSync(buffer);
 
         editor.scan(regex_if, scanResult => {
             // console.log('hello');
@@ -144,14 +134,4 @@ export default {
         // return (this.modalPanel.isVisible() ? this.modalPanel.hide() : this.modalPanel.show());
         // return this.modalPanel.hide();
     },
-
-    // list of regex queries for common mistakes
-
-    // activate only on supported langs?
-
-    // this'll be an action that finds an if statement with only 1 equal sign
-    //
-    scanEditorForCommonMistakes() {
-        console.log("Not implemented!");
-    }
 };
