+++ sample-files/sample_a.java
@@ -17,7 +17,7 @@ public class Main
             System.out.println("This is ok");
         }
 
-        if (i == 0 || i = 0)
+        if (i == 0 || i == 0)
         {
             System.out.println(
                 "This always evaluates to true. The above conditional should be identified as a code smell.");
